+++ RecastDemo/Source/CrowdTool.cpp
@@ -53,7 +53,7 @@ static bool isectSegAABB(const double* sp, const double* sq,
 	// For all three slabs
 	for (int i = 0; i < 3; i++)
 	{
-		if (fabsf(d[i]) < EPS)
+        if (fabs(d[i]) < EPS)
 		{
 			// Ray is parallel to slab. No hit if origin not within slab
 			if (sp[i] < amin[i] || sp[i] > amax[i])
